+++ layers/CMSSchema/packages/custompost-category-mutations/src/Module.php
@@ -17,6 +17,7 @@ class Module extends AbstractModule
         return [
             \PoPCMSSchema\CustomPostMutations\Module::class,
             \PoPCMSSchema\Categories\Module::class,
+            \PoPCMSSchema\TaxonomyMutations\Module::class,
         ];
     }