+++ layers/CMSSchema/packages/custompost-category-mutations/src/MutationResolvers/PayloadableSetCategoriesOnCustomPostMutationResolverTrait.php
@@ -7,6 +7,8 @@ namespace PoPCMSSchema\CustomPostCategoryMutations\MutationResolvers;
 use PoPCMSSchema\CustomPostCategoryMutations\FeedbackItemProviders\MutationErrorFeedbackItemProvider;
 use PoPCMSSchema\CustomPostCategoryMutations\ObjectModels\CategoryDoesNotExistErrorPayload;
 use PoPCMSSchema\CustomPostMutations\MutationResolvers\PayloadableCustomPostMutationResolverTrait;
+use PoPCMSSchema\TaxonomyMutations\ObjectModels\LoggedInUserHasNoAssigningTermsToTaxonomyCapabilityErrorPayload;
+use PoPCMSSchema\TaxonomyMutations\FeedbackItemProviders\MutationErrorFeedbackItemProvider as TaxonomyMutationErrorFeedbackItemProvider;
 use PoPSchema\SchemaCommons\ObjectModels\ErrorPayloadInterface;
 use PoP\ComponentModel\Feedback\ObjectTypeFieldResolutionFeedbackInterface;
 
@@ -30,6 +32,12 @@ trait PayloadableSetCategoriesOnCustomPostMutationResolverTrait
                 MutationErrorFeedbackItemProvider::class,
                 MutationErrorFeedbackItemProvider::E2,
             ] => new CategoryDoesNotExistErrorPayload(
+                $feedbackItemResolution->getMessage(),
+            ),
+            [
+                TaxonomyMutationErrorFeedbackItemProvider::class,
+                TaxonomyMutationErrorFeedbackItemProvider::E10,
+            ] => new LoggedInUserHasNoAssigningTermsToTaxonomyCapabilityErrorPayload(
                 $feedbackItemResolution->getMessage(),
             ),
             default => $this->upstreamCreateErrorPayloadFromObjectTypeFieldResolutionFeedback(